apps/backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
websockets==12.0
pyyaml==6.0.1
google-auth==2.28.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
python-dateutil==2.8.2
